requirements.txt
flask
requests
python-dotenv
gunicorn
flask
python-dotenv
requests
gunicorn
openai
